# docs/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerForge Pro API",
    "description": "CV-first career acceleration for South African remote work. Human-supervised: sensitive actions require explicit consent.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/profiles/{profile_id}/roles/recommend": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Recommend Roles",
        "operationId": "recommend_roles_api_v1_profiles__profile_id__roles_recommend_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleRecommendationOut"
                  },
                  "title": "Response Recommend Roles Api V1 Profiles  Profile Id  Roles Recommend Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles/{profile_id}/applications": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Create Application",
        "operationId": "create_application_api_v1_profiles__profile_id__applications_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "studio"
        ],
        "summary": "List Applications",
        "operationId": "list_applications_api_v1_profiles__profile_id__applications_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationOut"
                  },
                  "title": "Response List Applications Api V1 Profiles  Profile Id  Applications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{app_id}": {
      "get": {
        "tags": [
          "studio"
        ],
        "summary": "Get Application",
        "operationId": "get_application_api_v1_applications__app_id__get",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{app_id}/status": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Update Status",
        "operationId": "update_status_api_v1_applications__app_id__status_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{app_id}/tailor": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Tailor Application",
        "operationId": "tailor_application_api_v1_applications__app_id__tailor_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Tailor Application Api V1 Applications  App Id  Tailor Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{app_id}/cover-letter": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Create Cover Letter",
        "operationId": "create_cover_letter_api_v1_applications__app_id__cover_letter_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CoverLetterCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverLetterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{app_id}/videos": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Create Video Response",
        "operationId": "create_video_response_api_v1_applications__app_id__videos_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/videos/{video_id}/regenerate": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Regenerate Script",
        "description": "Re-run the script with new instructions; keeps version history.",
        "operationId": "regenerate_script_api_v1_videos__video_id__regenerate_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/videos/{video_id}/media": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Update Video Media",
        "description": "Client reports the recording state after upload (media storage lands with the studio UI).",
        "operationId": "update_video_media_api_v1_videos__video_id__media_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoMediaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles/{profile_id}/auto-pipeline": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Auto Pipeline",
        "description": "'Upload your CV and the program does the rest.'\n\nFor each job: pick the best master, tailor a job-specific CV, draft a\nhuman-sounding cover letter, and prepare a video-response script\n(when video consent exists). Everything lands as reviewable\napplications - nothing is sent or submitted.",
        "operationId": "auto_pipeline_api_v1_profiles__profile_id__auto_pipeline_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoPipelineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoPipelineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationCreate": {
        "properties": {
          "jd_id": {
            "type": "string",
            "title": "Jd Id"
          },
          "cv_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cv Version Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "jd_id"
        ],
        "title": "ApplicationCreate"
      },
...
      "ApplicationStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "pattern": "^(saved|ready|applied|phone_screen|interview|offer|rejected|archived)$",
            "title": "Status"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ApplicationStatusUpdate"
      },
...
      "CoverLetterCreate": {
        "properties": {
          "tone": {
            "type": "string",
            "pattern": "^(direct|warm|formal)$",
            "title": "Tone",
            "default": "direct"
          }
        },
        "type": "object",
        "title": "CoverLetterCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleRecommendationOut": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "match_pct": {
            "type": "number",
            "title": "Match Pct"
          },
          "matched": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "role",
          "match_pct",
          "matched",
          "missing",
          "reason"
        ],
        "title": "RoleRecommendationOut"
      },
...
      "VideoMediaUpdate": {
        "properties": {
          "media_status": {
            "type": "string",
            "pattern": "^(uploaded|ready)$",
            "title": "Media Status"
          }
        },
        "type": "object",
        "required": [
          "media_status"
        ],
        "title": "VideoMediaUpdate"
      },
...
    }
  }
}